requirements.txt
apache-airflow==2.9.0
black==23.3.0
fastapi==0.110.0
flask==2.2
mlflow==2.10.2
pandas==2.0.3
psycopg2-binary==2.9.7
pydantic==2.5.2
pydantic_core==2.14.5
pylint==2.17.3
requests==2.31.0
scikit-learn==1.3.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
uvicorn==0.27.1
uvloop==0.19.0
